backend/requirements.txt
django>=1.11,<1.12
psycopg2
django-filter==1.0.1
djangorestframework==3.5.3
djangorestframework-camel-case==0.2.0
markdown
gunicorn
pillow
django-cleanup

